+++ R/model-helpers.R
@@ -1,10 +1,11 @@
-#' Helpers for interacting with Julia model objects and functions
+#' Re-exported functions for interacting with Julia model objects
 #'
-#' @name jlme-model-helpers
+#' @name jlme-model-reexports
 #'
 #' @param x Julia model object
+#' @param ... Additional arguments passed to the Julia function
 #'
-#' @return An appropriate R object
+#' @return An appropriate R or Julia object
 #'
 #' @examplesIf check_julia_ok()
 #' \donttest{
@@ -18,11 +19,17 @@
 #' # `issingular()` reports whether model has singular fit
 #' issingular(x)
 #'
+#' # `likelihoodratiotest()` conducts a likelihood-ratio test between nested models
+#' likelihoodratiotest(
+#'   x,
+#'   jlmer(r2 ~ 1 + (1 | id), lme4::VerbAgg, family = "binomial")
+#' )
+#'
 #' stop_julia()
 #' }
 NULL
 
-#' @rdname jlme-model-helpers
+#' @rdname jlme-model-reexports
 #' @export
 propertynames <- function(x) {
   stopifnot(is_jl(x))
@@ -30,9 +37,20 @@ propertynames <- function(x) {
   sort(as.character(nm))
 }
 
-#' @rdname jlme-model-helpers
+#' @rdname jlme-model-reexports
 #' @export
 issingular <- function(x) {
   stopifnot(is_jl(x, "MixedModel"))
   JuliaConnectoR::juliaCall("MixedModels.issingular", x)
+}
+
+#' @rdname jlme-model-reexports
+#' @export
+likelihoodratiotest <- function(x, ...) {
+  stopifnot(is_jl(x))
+  model1 <- x
+  models_rest <- list(...)
+  all_models <- c(list(model1), models_rest)
+  fn <- JuliaConnectoR::juliaFun("MixedModels.likelihoodratiotest")
+  do.call(fn, all_models)
 }