+++ tests/testthat/test-contrasts.R
@@ -1,3 +1,5 @@
+stop_julia()
+
 test_that("custom contrast detection works", {
 
   x <- data.frame(v = c("a", "b", "c"))
@@ -21,9 +23,7 @@ test_that("formula construction works", {
 
   strsquish <- function(x) gsub("\\s", "", x)
 
-  expect_equal(
-    strsquish(construct_contrasts(x, "v", format = FALSE)),
-    strsquish('
+  expect_identical(strsquish(construct_contrasts(x, "v", format = FALSE)), strsquish('
       Dict(
         :v => HypothesisCoding(
           [ -1/2  1/2   0;
@@ -32,7 +32,6 @@ test_that("formula construction works", {
           labels=["helm1", "helm2"]
         )
       )
-    ')
-  )
+    '))
 
 })